change to double

+++ hw1/src/hdr.hpp
@@ -9,10 +9,10 @@ using namespace std;
 class DEBEVEC {
 
 public:
-  DEBEVEC(const vector<Mat>& pics, const vector<float>& etimes,
+  DEBEVEC(const vector<Mat>& pics, const vector<double>& etimes,
           const vector<Point>& points) 
     : _pics(pics), _etimes(etimes), _points(points) {};
-  void process(Mat&, float);
+  void process(Mat&, double);
 
 private:
   int W(int val) { return (val<=127 ? val+1 : 256-val); }
